// src/enums/config.key.code.enum.ts
export enum ConfigKeyCode {
  LIMIT_NO_LOSS_FREE = 'LIMIT_NO_LOSS_FREE',
  LIMIT_NO_LOSS_VIP = 'LIMIT_NO_LOSS_VIP',
  LIMIT_MUSIC_FREE = 'LIMIT_MUSIC_FREE',
  LIMIT_MUSIC_VIP = 'LIMIT_MUSIC_VIP',
  LIMIT_SOUND_FREE = 'LIMIT_SOUND_FREE',
  LIMIT_SOUND_VIP = 'LIMIT_SOUND_VIP',
  LIMIT_BOOKMARK_SOUND_ALBUM_FREE = 'LIMIT_BOOKMARK_SOUND_ALBUM_FREE',
  LIMIT_BOOKMARK_SOUND_ALBUM_VIP = 'LIMIT_BOOKMARK_SOUND_ALBUM_VIP',
  LIMIT_SOUND_CACHE_FREE = 'LIMIT_SOUND_CACHE_FREE',
  LIMIT_SOUND_CACHE_VIP = 'LIMIT_SOUND_CACHE_VIP',
  LIMIT_DOWNLOAD_FREE = 'LIMIT_DOWNLOAD_FREE',
  LIMIT_DOWNLOAD_VIP = 'LIMIT_DOWNLOAD_VIP',
  XMLY_FREE_STATUS = 'XMLY_FREE_STATUS',

  XMLY_VIP_COOKIE = 'XMLY_VIP_COOKIE',
  XMLY_VIP_BID = 'XMLY_VIP_BID',
  XMLY_VIP_CTN = 'XMLY_VIP_CTN',
}

export enum ConfigValueCode {
  DISABLED = 'DISABLED',
  ENABLED = 'ENABLED',
}